+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add JMasterDetail component [#559](https://github.com/nbbrd/java-desktop-util/issues/559)
+- Add JRangeSlider component [#560](https://github.com/nbbrd/java-desktop-util/issues/560)
 
 ## [2.6.2] - 2026-04-09